Move & Rename API `process_pending_scroll_events`.
